+++ Scripts/languages/php.js
@@ -183,7 +183,7 @@ class PHPParser extends LanguageParser {
         let typed = match.groups.tag;
         
         tags.forEach(tag => {
-            if (tag[0].substr(0, typed.length) === typed) {
+            if (tag[0].includes(typed)) {
                 matches.push(tag);
             }
         });