+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 0.5
+
+* Comment extension _(experimental feature)_  
+  Pressing return inside a docblock will try to insert a leading asterisk and maintain indentation. This is very hacky code, which constantly tracks cursor position and docblock contents. This **may** slow down your editor, which is why this is an experimental feature and disabled by default.
+* minor fixes
+
+
 ## Version 0.4
 
 * Custom tags for header blocks (global and per workspace)